test(ui): Playwright e2e for trace flow (T5)
Mock-driven e2e covering: list view → SqliteBanner + PipelineHealthPanel
visible + trace rows render → detail page (placeholder route under
output:export) renders Tree + Waterfall → SpanDetailRail tab switching.

Splits the dynamic detail route into a server-side `page.tsx`
(generateStaticParams placeholder, required by Next.js static export)
+ a `TraceDetailClient.tsx` client component holding the actual
3-panel UI. The unit test imports the client component directly.
